WordcloudProject/requirements.txt
janome.tokenizer
wordcloud
matplotlib
os